To make the samplebot more flexible. I have implemented the Strategy pattern for the actions that the bot can take.
Strategy Interface has two signatures
- getPreFlop
and
- getPostFlop

The init() method of the sample player now reflectively creates a
strategy for itself. The strategy is read as a string from the bots
preference .pd file and the Class.newInstance() method is called to
create an instance of the strategy for the SamplePlayer to use.